Fix tab navigation and screen mappings
